Added USART defines for all pins
Serial communication was successful by changing the Fan5 Tach pin to
serial. This was the pin configured as serial in R1.2. This commit
initialises all the serial pins available. Trial and error will be used
to find the right pin define for R2.0. Datasheet does not match the code
pin defines which is why this is being done.