Electric power input changes, cleaned up other inputs
- Electric power input changes
-- Changed FC Voltage, FC Current, Cap Voltage, Cap Current conversion
constants defined in ADC conversion constants. The voltage constants are
from a calibration procedure outlined in an Excel document on the Systems
Drive "FCVoltandCapVoltCalib.xlsx". The current constants are from a V=IR
calculation. R is the resistance of the shunt resistor (0.001 ohms)
multiplied by the gain.
-- get_CAPCURR() function is now uncommented. The output has been added to
the datalogging.c file

- Cleaned up other inputs
-- Ambient temperature input code has now been removed. This was never
used and the inputs were removed from the schematic in R2.0 of the AVR
FCC.
-- Tank Pressure input code has also been removed. This was also never
used and the input doesn't exist on the AVR FCC R2.0 anymore.