Added Manual Depress mode, LED struct architecture
This version of the code contains a new run mode called manual
depressurize. This new run mode runs for 500 ms does the following:
	- Turns on status LED 2
	- Opens all of the relays and the purge valve
	- Turns on LED 0 to indicate that the valve is open

After 500 ms:
	- It closes all of the relays and the purge valve
	- Turns off LED 0 and LED 2
	- Goes to the alarm state

In order to activate the run mode, BUTTON 1 should be held down
for five (5) seconds. After this button is held down for five (5)
seconds, LED 3 begins to flash until the CPU goes into the alarm
state.

Furthermore, this code implements a new architecture for working
with LEDs. A structure template has been created to work the LED.
This makes it easy to access and store information for each LED,
and also allows for passing the information about the LEDs into
user-defined functions. The LED structure contains the following
data:
	- Pin name
	- The current state of the LED (1 for on, 0 for off)
	- A timer used to make the LED blink

The following two functions
have been written to accommodate this archtecture:

changeLEDstate:
	- This function updates the state of the LED (on/off)
	- It has no return value
	- It takes a pointer to an LED structure as an argument
	- It also takes an integer that determines the new mode
	  as an argument

	- Prototype: void changeLEDstate(struct LED *, int);

LEDblink:
	- This function causes the LED to change state, and
	  causes therefore causes it blink if it is used in
	  a loop
	- It has no return value
	- It takes a pointer to an LED structure as an argument
	- It also takes a long that determines the delay between
	  blinks

	- Prototype: void LEDblink(struct LED *, long);

To be Tested:
	- The code appears to be cycling through the manual
	  depressurize run mode correctly, however the outputs
	  need to be tested
	- The LED struct code needs to be tested to ensure that
	  it is controlling the LEDs as desired. LED 3 should
	  begin to blink as soon as BUTTON 1 is held down for
	  five (5) seconds so the code can be tested that way